requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-amazon  #this provider allows connection to s3 bucket
apache-airflow-providers-postgres #this provider allows connection to redshift(postgres)
